Move WebSocketViewModel to viwModels root directory
